Link groups instead of copying values
Using `syntax['A'] = syntax['B']` assigns the values to group B to group
A but it doesn't link them.  This means modifying B through the
`highlight` command later on does not affect A.

By using `highlight link` instead of assigning, the connection between A
and B is preserved so all modifications made to B will also affect A.
This makes customisations easier, as, for example, modifying `Delimiter`
will also affect linked groups like `TSPunctDelimiter`,
`TSPunctBracket`, etc.